+++ archon-ui-main/src/components/settings/CodeExtractionSettings.tsx
@@ -3,7 +3,7 @@ import { Code, Check, Save, Loader } from 'lucide-react';
 import { Card } from '../ui/Card';
 import { Input } from '../ui/Input';
 import { Button } from '../ui/Button';
-import { useToast } from '../../features/ui/hooks/useToast';
+import { useToast } from '../../features/shared/hooks/useToast';
 import { credentialsService } from '../../services/credentialsService';
 
 interface CodeExtractionSettingsProps {
