+++ archon-ui-main/src/features/mcp/components/McpConfigSection.tsx
@@ -1,7 +1,7 @@
 import { Copy, ExternalLink } from "lucide-react";
 import type React from "react";
 import { useState } from "react";
-import { useToast } from "../../ui/hooks";
+import { useToast } from "@/features/shared/hooks";
 import { Button, cn, glassmorphism, Tabs, TabsContent, TabsList, TabsTrigger } from "../../ui/primitives";
 import type { McpServerConfig, McpServerStatus, SupportedIDE } from "../types";
 import { copyToClipboard } from "../../shared/utils/clipboard";
