+++ archon-ui-main/src/features/projects/tasks/hooks/tests/useTaskQueries.test.ts
@@ -20,14 +20,14 @@ vi.mock("../../services", () => ({
 const showToastMock = vi.fn();
 
 // Mock the toast hook
-vi.mock("../../../../ui/hooks/useToast", () => ({
+vi.mock("../../../../shared/hooks/useToast", () => ({
   useToast: () => ({
     showToast: showToastMock,
   }),
 }));
 
 // Mock smart polling
-vi.mock("../../../../ui/hooks", () => ({
+vi.mock("../../../../shared/hooks", () => ({
   useSmartPolling: () => ({
     refetchInterval: 5000,
     isPaused: false,
