+++ archon-ui-main/src/features/settings/version/hooks/useVersionQueries.ts
@@ -4,7 +4,7 @@
 
 import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
 import { STALE_TIMES } from "@/features/shared/queryPatterns";
-import { useSmartPolling } from "@/features/ui/hooks/useSmartPolling";
+import { useSmartPolling } from "@/features/shared/hooks/useSmartPolling";
 import { versionService } from "../services/versionService";
 import type { VersionCheckResponse } from "../types";
