+++ archon-ui-main/src/features/shared/hooks/useToast.ts
@@ -1,6 +1,6 @@
 import { AlertCircle, CheckCircle, Info, XCircle } from "lucide-react";
 import { createContext, useCallback, useContext, useEffect, useRef, useState } from "react";
-import { createOptimisticId } from "../../shared/optimistic";
+import { createOptimisticId } from "../optimistic";
 
 // Toast types
 interface Toast {
